Extracting ULRA specific queries
While it is possible this method would exist across multiple work
areas, I'd like to keep the logic separated for the time being.
