+++ app/repositories/sipity/queries/additional_attribute_queries.rb
@@ -26,15 +26,6 @@ module Sipity
         returning_value = returning_value.first if cardinality == :one || cardinality == 1
         returning_value
       end
-
-      def available_supervising_semester_for(ending_year: Time.zone.today.year, **)
-        (-3..0).each_with_object([]) do |year_delta, values|
-          year = ending_year.to_i + year_delta
-          ['Spring', 'Summer', 'Fall'].each do |season|
-            values << "#{season} #{year}"
-          end
-        end
-      end
     end
   end
 end
