+++ spec/repositories/sipity/queries/additional_attribute_queries_spec.rb
@@ -23,16 +23,4 @@ RSpec.describe Sipity::Queries::AdditionalAttributeQueries, type: :isolated_repo
         to eq(['Title is Chicken', 'Title is Egg'])
     end
   end
-  context '#available_supervising_semester_for' do
-    it 'will return an array of semesters' do
-      expect(test_repository.available_supervising_semester_for(ending_year: 2015, work: work)).to eq(
-        [
-          "Spring 2012", "Summer 2012", "Fall 2012",
-          "Spring 2013", "Summer 2013", "Fall 2013",
-          "Spring 2014", "Summer 2014", "Fall 2014",
-          "Spring 2015", "Summer 2015", "Fall 2015"
-        ]
-      )
-    end
-  end
 end
